Add a presenter for the list of all users, and employee summary.
