+++ src/DayView.java
@@ -1,5 +1,6 @@
 
 import Entities.Shift;
+import FrameworksAndDrivers.Page;
 
 import javax.swing.*;
 import java.awt.*;
