+++ src/FrameworksAndDrivers/ManageEmployeesGUI.java
@@ -8,7 +8,7 @@ import java.awt.event.ActionListener;
 
 public class ManageEmployeesGUI implements ActionListener, Page {
 
-
+  private int viewerID;
   private JFrame frame = new JFrame();
   private JButton addEmployee = new JButton("Add Entities.Employee");
 
@@ -18,7 +18,7 @@ public class ManageEmployeesGUI implements ActionListener, Page {
 
 
 public ManageEmployeesGUI(){
-
+  //Create the UI by combining the title and button components
   frame.setSize(600, 600);
   frame.setVisible(true);
   this.addTitle();
@@ -30,6 +30,7 @@ public ManageEmployeesGUI(){
 
 @Override
 public void actionPerformed(ActionEvent e){
+  //Based on which button is clicked, we wish to redirect to a particular page.
   Object source = e.getSource();
     if (source.equals(addEmployee)) {
       new AddEmployeeGUI();
@@ -38,7 +39,7 @@ public void actionPerformed(ActionEvent e){
       new CompleteUserListGUI();
       frame.dispose();
     } else if (source.equals(employeeSummary)) {
-      new EmployeeSummaryGUI();
+      new EmployeeSummaryGUI(viewerID);
       frame.dispose();
     }
 
@@ -51,6 +52,7 @@ public void actionPerformed(ActionEvent e){
 
   @Override
   public void addContent() {
+    //Make the buttons respond to a click, and add them to the button panel.
     this.buttonsPanel.setLayout(new GridLayout(3, 1));
     addEmployee.addActionListener(this);
     completeEmployeeList.addActionListener(this);
@@ -60,6 +62,27 @@ public void actionPerformed(ActionEvent e){
     this.buttonsPanel.add(employeeSummary);
     this.buttonsPanel.setSize(100,300);
     frame.add(buttonsPanel);
+  }
+
+  @Override
+  public void setUser(int user) {
+  /*We need this to keep track of who is viewing the page.*/
+    this.viewerID = user;
+  }
+
+  @Override
+  public void dispose() {
+
+  }
+
+  @Override
+  public void addHomeButton() {
+
+  }
+
+  @Override
+  public void update() {
+
   }
 }
