+++ src/FrameworksAndDrivers/Page.java
@@ -1,9 +1,12 @@
+package FrameworksAndDrivers;
+
 import javax.swing.*;
 
 public interface Page {
 
     JButton back = new JButton("Back");
     JPanel backPanel = new JPanel();
+
     public void addTitle();
     public void addContent();
     public void setUser(int user);
