+++ src/HomePage.java
@@ -1,13 +1,14 @@
 import Entities.HR;
 import FrameworksAndDrivers.AddHRGUI;
 import FrameworksAndDrivers.ManageEmployeesGUI;
+import FrameworksAndDrivers.Page;
 
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class HomePage implements ActionListener, Page{
+public class HomePage implements ActionListener, Page {
 
     private JFrame frame = new JFrame();
