+++ src/InterfaceAdapters/RequestFormPresenter.java
@@ -1,6 +1,7 @@
 package InterfaceAdapters;
 
 import Entities.NotificationRequest;
+import FrameworksAndDrivers.Page;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
