+++ src/LoginValidator.java
@@ -1,12 +1,13 @@
 import Entities.Employee;
 import Entities.User;
+import InterfaceAdapters.UserInteractor;
 
 import java.util.ArrayList;
 
 public class LoginValidator {
     private int employeeID;
     private char[] password;
-    private EmployeeDataBaseInteractor empDB = new EmployeeDataBaseInteractor();
+    private UserInteractor empDB = new UserInteractor();
 
     public User validateCredentials(int empID, char[] pwd){
         ArrayList<User> allActiveUsers = empDB.readData();
